requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
astroid==2.14.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
DateTime==5.1
dill==0.3.6
docutils==0.18.1
flake8==6.0.0
flake8-for-pycharm==0.4.1
fonttools==4.38.0
frozenlist==1.3.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
hydra-core==1.3.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-loader==0.1rc2
lazy-object-proxy==1.9.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nibabel==5.0.0
nilearn==0.10.0
numpy==1.24.1
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.0
pandas==1.5.3
pandas-flavor==0.3.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.16.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
snowballstemmer==2.2.0
sphinx==6.1.3
sphinx-autodoc-typehints==1.22
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jquery==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.1
torch-geometric==2.2.0
torch-scatter==2.1.0+pt113cpu
torch-sparse==0.6.16+pt113cpu
torch-spline-conv==1.2.1+pt113cpu
tqdm==4.64.1
typing==3.7.4.3
typing-extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
xarray==2023.1.0
yarl==1.8.2
zipp==3.12.0
zope.interface==5.5.2
